refactor: split browser manager modules
